Refractoring

+++ Chess3.0/src/Hash.h
@@ -7,7 +7,7 @@ struct entry {
     uint64_t key;
     int depth;
     int flags;
-    int value;
+    float value;
 };
 
 void gen_random(uint64_t(&random_key)[769]);
@@ -17,8 +17,8 @@ public:
 	entry* Table;
 	TranspositionTable(uint64_t size);
 
-    void store(int depth, int val, int flag, uint64_t key);
-    int probe_hash(int depth, int alpha, int beta, uint64_t key);
+    void store(int depth, float val, int flag, uint64_t key);
+    int probe_hash(int depth, float alpha, float beta, uint64_t key);
 };
